Implement cpu load

+++ app_analysis/inc/app_analysis_const.h
@@ -42,7 +42,7 @@ typedef struct PROCESS_INFO_TAG
     uint32_t num_threads;
     uint32_t memory_size;
     uint32_t handle_cnt;
-    uint32_t cpu_load;
+    float cpu_load;
 } PROCESS_INFO;
 
 typedef struct NETWORK_INFO_TAG
